se avanza con endpoints

+++ src/main/java/com/datamusic/datamusic/persistence/GeneroRepository.java
@@ -4,11 +4,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Repository;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.datamusic.datamusic.domain.Gender;
 import com.datamusic.datamusic.domain.repository.GenderRepository;
+import com.datamusic.datamusic.persistence.crud.AlbumCrudRepository;
 import com.datamusic.datamusic.persistence.crud.GeneroCrudRepository;
+import com.datamusic.datamusic.persistence.entity.AlbumEntity;
 import com.datamusic.datamusic.persistence.entity.Genero;
 import com.datamusic.datamusic.persistence.mapper.GenderMapper;
 
@@ -17,6 +21,9 @@ public class GeneroRepository implements GenderRepository {
 
     @Autowired
     private GeneroCrudRepository generoCrudRepository;
+
+    @Autowired
+    private AlbumCrudRepository albumCrudRepository;
 
     @Autowired
     private GenderMapper mapper;
@@ -39,8 +46,19 @@ public class GeneroRepository implements GenderRepository {
     }
 
     @Override
-    public void delete(Long genderId) {
+    public boolean delete(Long genderId) {
+        //verificar si existe en la tabla albums , albums con ese id genero , si lo hay no permitir eliminar 
+        // genero por violacion de llave foranea 
+
+        List<AlbumEntity> albumsXgenero=albumCrudRepository.findByIdGenero(genderId);
+        if (!albumsXgenero.isEmpty()) {
+            Optional<Gender> gender=getGender(genderId);
+            String genderName = gender.map(Gender::getName).orElse("");
+            throw new ResponseStatusException(
+                HttpStatus.CONFLICT, "No se puede eliminar el género " + genderName + " porque tiene álbumes asociados");
+        }
         generoCrudRepository.deleteById(genderId);
+        return true;
     }
 
 }
